+++ src/syntax/menai/menai_lexer.py
@@ -464,6 +464,6 @@ class MenaiLexer(Lexer):
             True if the value is a special form, False otherwise
         """
         special_forms = {
-            'and', 'if', 'let', 'let*', 'letrec', 'lambda', 'or', 'quote', 'match', 'import', 'trace', 'struct', 'apply'
+            'and', 'if', 'let', 'let*', 'letrec', 'lambda', 'or', 'quote', 'match', 'import', 'export', 'struct', 'apply', '::'
         }
         return value.lower() in special_forms